refactor(shared): split misnamed auth.ts into answers.ts + streaks.ts
