+++ packages/shared/src/answers.ts
@@ -15,15 +15,4 @@ export const AnswerQuestionResponseSchema = z.object({
   xpAwarded: z.number().int().min(0),
 });
 
-export type AnswerQuestionResponse = z.infer<
-  typeof AnswerQuestionResponseSchema
->;
-
-/** GET /streaks — response */
-export const StreakResponseSchema = z.object({
-  currentStreak: z.number().int().min(0),
-  longestStreak: z.number().int().min(0),
-  totalSessions: z.number().int().min(0),
-});
-
-export type StreakResponse = z.infer<typeof StreakResponseSchema>;
+export type AnswerQuestionResponse = z.infer<typeof AnswerQuestionResponseSchema>;
